rearrange methods from post_upgrade over to init
